remove: time selector 지움 (FE-128)
